perf(ui): keep catalog section rendering out of startup bundle
