+++ ui/src/components/app-sidebar.ts
@@ -25,6 +25,7 @@ import {
   renderAppSidebarPluginTabEntry,
   renderAppSidebarZoneEntry,
 } from "./app-sidebar-render.ts";
+import type { SessionCatalogGroupsRenderer } from "./app-sidebar-session-catalog-render.ts";
 import type { CatalogSessionMenuRequest } from "./app-sidebar-session-catalogs.ts";
 import { renderSessionList } from "./app-sidebar-session-list-render.ts";
 import type {
@@ -73,6 +74,19 @@ class AppSidebar extends AppSidebarSessionNavigationElement implements SessionLi
   private narration: SidebarSessionNarrationController | null = null;
   private narrationLoad: Promise<void> | null = null;
   private readonly narrationSubscriptions = this.createNarrationSubscriptions();
+
+  // Catalog rows are non-startup content. Load their renderer through the same
+  // idle boundary as other sidebar chrome, then repaint when the chunk arrives.
+  private catalogRenderer: SessionCatalogGroupsRenderer | null = null;
+  private readonly catalogRendererImport = createIdleImport(
+    () => import("./app-sidebar-session-catalog-render.ts"),
+    (module) => {
+      this.catalogRenderer = module.renderSessionCatalogGroups;
+      if (this.isConnected) {
+        this.requestUpdate();
+      }
+    },
+  );
 
   @state() catalogProjectGrouping = loadStoredSidebarCatalogGrouping();
 
@@ -123,6 +137,7 @@ class AppSidebar extends AppSidebarSessionNavigationElement implements SessionLi
 
   override disconnectedCallback() {
     this.narration?.disconnect();
+    this.catalogRendererImport.dispose();
     super.disconnectedCallback();
   }
 
@@ -206,6 +221,7 @@ class AppSidebar extends AppSidebarSessionNavigationElement implements SessionLi
     // The decorative pet's large module stays out of startup and upgrades in place.
     // Its first visit is at least 15 seconds after load, so idle loading cannot miss one.
     sidebarChromeImport.schedule();
+    this.catalogRendererImport.schedule();
   }
 
   protected override firstUpdated() {
@@ -285,6 +301,10 @@ class AppSidebar extends AppSidebarSessionNavigationElement implements SessionLi
     this.sessionData.dismissSessionMutationError();
   }
 
+  preloadCatalogRenderer() {
+    return this.catalogRendererImport.load();
+  }
+
   setCatalogProjectGrouping(next: CatalogProjectGrouping): void {
     storeSidebarCatalogGrouping(next);
     this.catalogProjectGrouping = next;
@@ -318,10 +338,18 @@ class AppSidebar extends AppSidebarSessionNavigationElement implements SessionLi
       }
     }
     const { sections } = this.zonedVisibleSections(visibleSessions);
+    if (
+      !this.catalogRenderer &&
+      (this.sessionData.sessionCatalogs.length > 0 ||
+        this.sessionData.sessionCatalogRefreshStatus.error !== null)
+    ) {
+      void this.preloadCatalogRenderer().catch(() => undefined);
+    }
     return renderSessionList({
       host: this,
       empty: visibleSessions.length === 0,
       sections,
+      catalogRenderer: this.catalogRenderer,
       showDraft:
         Boolean(this.draftSessionAgentId) &&
         normalizeAgentId(this.draftSessionAgentId) === expandedAgentId,
